+++ src/julia.h
@@ -827,7 +827,7 @@ DLLEXPORT uv_loop_t *jl_global_event_loop();
 DLLEXPORT uv_pipe_t *jl_make_pipe(int writable, int julia_only, jl_value_t *julia_struct);
 DLLEXPORT void jl_close_uv(uv_handle_t *handle);
 
-DLLEXPORT int16_t jl_start_reading(uv_stream_t *handle);
+DLLEXPORT int32_t jl_start_reading(uv_stream_t *handle);
 
 DLLEXPORT void jl_callback(void *callback);
